requirements.txt
opencv-python == 4.5.5.62
numpy == 1.22.*
matplotlib
coloredlogs
opencv-contrib-python == 4.5.5.64
pillow == 9.0.*
pyyaml
tqdm
numba